chore(tp): moves tp cmd logic to wrapper
reason for this change is to decouple develop command with tp-specific operations
